Fixed policy row redirect, renaming and remove it unused code
