+++ public/components/agents/sca/inventory/index.ts
@@ -1,4 +1,4 @@
 export * from './api-request'
-export * from './policies-table'
+export * from './agent-policies-table'
 export * from './lib';
 export * from './checks-table';
